requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dateutil>=2.8.2
Werkzeug>=3.0.0
gunicorn>=21.2.0
